added some treatments of not found exceptions

+++ src/main/java/com/hackaton/hackatondonatessystem/services/CauseService.java
@@ -101,8 +101,13 @@ public class CauseService {
         return companiesDTO;
     }
 
-    public void delete(Long id){
-        repository.deleteById(id);
+    public void delete(Long id) throws NotFoundException {
+        if(repository.existsById(id)){
+            repository.deleteById(id);
+        } else{
+            throw new NotFoundException("Cause not found");
+        }
+
     }
 
     private CauseDTO convertCausesToDTO(Cause cause){
