+++ src/main/java/com/hackaton/hackatondonatessystem/services/CompanyService.java
@@ -91,8 +91,12 @@ public class CompanyService {
         return companyDTO;
     }
 
-    public void delete(Long id){
-        repository.deleteById(id);
+    public void delete(Long id) throws NotFoundException {
+        if(repository.existsById(id)){
+            repository.deleteById(id);
+        } else{
+            throw new NotFoundException("Cause not found");
+        }
     }
