+++ src/main/java/com/hackaton/hackatondonatessystem/services/UserService.java
@@ -81,8 +81,12 @@ public class UserService{
         return memberDTO;
     }
 
-    public void delete(Long id){
-        repository.deleteById(id);
+    public void delete(Long id) throws NotFoundException {
+        if(repository.existsById(id)){
+            repository.deleteById(id);
+        } else{
+            throw new NotFoundException("Cause not found");
+        }
     }
